Progress generally, work-around a weird bug
Revise v.zsh add DEBUG and insert superflous quotes in sed command soas
not to match itself when editing itself

Add check and fix for curpos in --msgify and output a \n if it's not at
column 1.

Revise :assign to avoid a weird hang by read when the input has stderr
redirected to /dev/null. This just started happening. Not sure why.

Revise :get:current:branch to NOT report error conditions other than to
set $?.

Add run-help support to zsh.rc. I can't understand why the default setup
is for help to do as little as possible and duplicate (by calling) man.